[dingo-calcite] Add float that supports precision

+++ dingo-calcite/src/main/java/io/dingodb/calcite/type/converter/RexLiteralConverter.java
@@ -73,8 +73,8 @@ public class RexLiteralConverter implements DataConverter {
     }
 
     @Override
-    public Double convertFloatFrom(@NonNull Object value) {
-        return ((BigDecimal) value).doubleValue();
+    public Float convertFloatFrom(@NonNull Object value) {
+        return ((BigDecimal) value).floatValue();
     }
 
     @Override
